updata:回收站文件恢复、删除

+++ networkDisk/networkDisk-filestorage/src/main/java/com/networkDisk/filestorage/domain/NetworkdiskUserFilestorage.java
@@ -50,7 +50,7 @@ public class NetworkdiskUserFilestorage extends BaseEntity {
     /**
      * 文件后缀名
      */
-    private Long fileSize;
+    private Double fileSize;
     /**
      * URL地址
      */
