+++ networkDisk/networkDisk-system/src/main/java/com/networkDisk/system/domain/SysOss.java
@@ -40,7 +40,7 @@ public class SysOss extends BaseEntity {
     /**
      * 文件大小
      */
-    private Long fileSize;
+    private Double fileSize;
 
     /**
      * URL地址
